perf: DH-23586: Address review comments.
- Evaluate the sparse-overhead check at the start of onUpdate against the
  validated table's already-updated row set, migrating before the update
  records any values at newly scattered keys; previously a single
  scattering update could allocate unbounded sparse structure before the
  end-of-cycle check ran (and the migration mutated state after
  downstream listeners had been notified).
- Skip the redirection for zero-column validators, which have no expected
  values to compact.
- Merge the new tests into the existing TestTableUpdateValidator instead
  of a new test class, and use @VisibleForTesting on isRedirectionUsed.
